Creating new module with variables
A new module called mod_vars was created. It contains all the variables
read from the input file and also check of the variables. Now, I can
source this module in all following modules a I don't need to pass ngrid
and dx in all the functions.

Tested on harmonic oscillator of ranks 1-3 and also first excited state
of 1D HO.