Correcting imaginary propagation! Until now phase rotation present!
This commit corrects the imaginary time propagation, which multiplied
the wave function by (1+1i) every step which caused rotation and
increase of norm. Renormalization corrected it but still I had norm
rotation. It did not influence the energy or norm of wavefunction so it
was not possible to spot it by watching these quantities.

The propagation is now correct, yet it still has problems by giving the
wave function a stationary phase. To be corrected!