Changing read.F90 module to vars.F90 and moving parts to init.F90
I changed read.F90 module so that it only reads the variables and
checks if they are ok. It then contains all the variables as global. All
the initializing procedures in there were moved to init.F90.

I also added new version of prepare_potential which know allows cubic
spline as interpolation and also allows to convolute data with Gaussian.