Removing back FFT from energy evaluation of <T>
I've removed the back FFT from calculations of energy for IT dynamics to
speed up the calculation. There is no need to transform psi*p back to
coordinate space and then doing the expectation value. I rather
calculated the <T> in the momentum space.

WARNING: Bracket function used in the <T> calculation uses dx, dy, dz
for the integration, not dpx, dpy, dpz!!! This will naturally cancel out
by normalization but it is worth to know about it.