Exact factorization (#16)
1D exact factorization module added to QDyn. The implementation is based on code by Federica Agostini but improved and debugged. The implementation was tested against her code, test suite was created.

* Progress on exact factorization.

* Added vars.F90

* First set of EF modifications
Preparations of the code to calculate EF quantities

* First running test version

Everything but GD-TDPES is now stable and calculated, yet it seems that the derivative dC is making some problems.
Plotting of the EF quantities is also partially prepared.

* New python scripts, mainly preparing analytic potentials (#17)

* Updating outputs in tests

* Fixed version where GD-TDPES should be more or less correct.

* Adding qdyn_analyze.py with reading functions

* Adding EF test to the test suite

* New progress with implementation of GD-TDPES

The GD-TDPES is implemented and calculation of all the components too. However, comparing to Fede's code, I observe very much different results. The code will be checked and bug discovered in the next commits.

* Correction of the sign in the last term of GD-TDPES

This is not fixing the bug reported previously, I just forgot to revert change to Fede's code.

* Getting to fixes to GD-TDPES yet still not correct

I got the right shape but I have a wrong sign of GD-TDPES compared to Fede.

* Simplifying implementation of GD-TDPES and getting rid of imaginary values

Reshuffling the GD-TDPES equation to end up with simpler and real formula which avoids problems with imaginary numbers.

* Resolving momentum calculation for nuclear phase/vector potential and fixing problems when interaction is off

Fixing the inconsistency in vecpotS0_1d(). Now I go correctly with iFFT --> -ipx*wf --> FFT. Copying from Fede, I didn't do the math properly to realize that in iFFT/FFT procedure I need minus. Fede uses equivalent FFT --> ipx*wf --> iFFT.

I fixed the bug in calc_gitdpes_1d() when interaction with elmag field was off.

* Formatting qdyn.F90 and removing todo from qdyn_analyze.py

No effective change to the code appeared, just formatting.

* Fixing the problems with large GD-TDPES at first step

The problem was in wf derivative since I was not saving the initial wave function to the history. The first wf I saved was after step 1 which left 0 in the history and made largde derivative.

* GD-TDPES fixed, matches reference now!!

The opposite sign in GD-TDPES was due to reverse order of indexes I had in my time derivatives using history arrays. The fix now turned the sign the right way and I could compare to Fede's code (after debugging it).

* Finishing EF tests and running the whole test suite

gd-tdpes_ref.dat was added to the EF test and the whole test suite was run to get the new output formatting uploaded.

* Improving exact_factorization.py